+++ Entity/UserManager.php
@@ -47,7 +47,7 @@ class UserManager extends BaseUserManager
     /**
      * {@inheritDoc}
      */
-    public function deleteUser(UserInterface $user)
+    protected function doDeleteUser(UserInterface $user)
     {
         $this->em->remove($user);
         $this->em->flush();
@@ -91,7 +91,7 @@ class UserManager extends BaseUserManager
      * @param UserInterface $user
      * @param Boolean       $andFlush Whether to flush the changes (default true)
      */
-    public function updateUser(UserInterface $user, $andFlush = true)
+    protected function doUpdateUser(UserInterface $user, $andFlush = true)
     {
         $this->updateCanonicalFields($user);
         $this->updatePassword($user);